+++ client/web/src/app/modules/platform/interaction/customer-info/customer-info.component.ts
@@ -5,6 +5,7 @@ import {
   ICIIcons,
 } from '../../../../core/interfaces/icidata.model';
 import { DataService } from 'src/app/core/services/data.service';
+import { Customer } from 'src/app/core/interfaces/customer.model';
 
 @Component({
   selector: 'app-customer-info',
@@ -15,7 +16,7 @@ export class CustomerInfoComponent {
   statusTypes = ICI_STATUS;
   icons: Array<ICIIcons> = [];
 
-  @Input() info!: ICIData; // TODO: set correct model type after API available
+  @Input() info!: Customer; // TODO: set correct model type after API available
   @Output() handleClick = new EventEmitter();
 
   constructor(private dataService: DataService) {}
@@ -28,8 +29,8 @@ export class CustomerInfoComponent {
     this.icons = this.dataService.icons;
   }
   isIconEnabled(icon: ICIIcons): boolean {
-    if (this.info.customerInfo.additionalInfo.icons) {
-      return this.info.customerInfo.additionalInfo.icons.includes(icon.id);
+    if (this.info.additionalInfo?.icons) {
+      return this.info.additionalInfo?.icons.includes(icon.id);
     }
     return false;
   }